requirements.txt
apache-airflow==2.6.1
pandas==1.5.3
pydantic==1.10.13
psycopg2-binary==2.9.3  # For PostgreSQL support
requests==2.28.1        # For HTTP requests
apache-airflow-providers-postgres==2.0.1  # Airflow provider for PostgreSQL
apache-airflow-providers-http==2.0.1 # Airflow provider for HTTP operations
flask-session==0.4.0 
